+++ DuoArtifacts/I18n.cs
@@ -105,7 +105,7 @@ internal static class I18n
 	public static readonly string BooksMaxArtifactTooltip = "Whenever you <c=cardtrait>exhaust</c> a card, gain 1 <c=status>SHARD</c>.";
 
 	public static readonly string BooksPeriArtifactName = "Shardblade Infuser";
-	public static readonly string BooksPeriArtifactTooltip = "Whenever you play an <c=action>attack</c>, lose 1 <c=status>SHARD</c>: the <c=action>attack</c> deals 1 more damage.";
+	public static readonly string BooksPeriArtifactTooltip = "Whenever you play an <c=action>attack</c> card while having 3 <c=status>SHARD</c>, all <c=action>attacks</c> of that card deal 1 more damage.";
 
 	public static readonly string BooksRiggsArtifactName = "Fleetfoot Resonator";
 	public static readonly string BooksRiggsArtifactTooltip = "Gain 1 <c=status>HERMES BOOTS</c> for each 3 <c=status>SHARD</c> you have each turn.";